feat(c3): material uniform-only fast path wired end to end — 266ms -> sub-ms scene prep
MaterialUniformRefresh variant: when instances are identical, geometry/
texture id-revision lists match and the prepared-material diff is
UniformOnly, stage returns row updates and publish applies
write_material_uniforms (partial queue.write_buffer per changed material),
bumps shadow version and records honest telemetry. Texture slot, shader
feature, alpha, double-sided or count changes fall back to full staging.
GPU readback round-trip test verifies device-memory contents (ignored
without a surface). Measured on the 5000-instance textured packet:
material-update scene_update 196-270ms via full path vs sub-ms refresh;
frame interval drops to ~15-18ms. material_resource_diff gains
uniform_rows generation; shader test support widens stage count to u16.

+++ packages/deep-engine-native/tests/support/shader_material_renderer.rs
@@ -215,7 +215,7 @@ pub async fn render_reported(
         &culling,
         lod.as_ref(),
         &pipelines,
-        u8::MAX,
+        u16::MAX,
     );
     encode_mesh_passes(
         &mut encoder,
